requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
docopt==0.6.2
idna==2.10
Js2Py==0.70
MouseInfo==0.1.3
numpy==1.19.3
opencv-python==4.4.0.44
packaging==20.4
Pillow==8.0.1
pip==20.2.4
pipwin==0.5.0
PyAudio==0.2.11
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
setuptools==49.2.1
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tzlocal==2.1
urllib3==1.25.11